requirements.txt
flask
requests
flask-cors
google-genai
python-dotenv
